Improved test coverage
Wrote some extra test cases to check that exceptions are raised when
invalid inputs are given to both the data and annotation modules

More specifically, added test for the newly added check valid path for
both valid states and for the excpetions raised. Added test case for
encode anotation and open json exceptions!

Closes #123, #124